pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ctx-modules"
version = "0.1.0"
description = "Context module system for RAG and AI coding tools"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5",
]

[project.optional-dependencies]
extractors = [
    "pymupdf>=1.23",
    "python-pptx>=0.6",
    "markdownify>=0.11",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
ctx = "ctx.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
